requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.11.1
gunicorn==20.0.4
importlib-metadata==1.7.0
kombu==4.6.11
psycopg2-binary==2.8.6
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
